Implement a video mode.
The regular firmware is designed to be easy to use, but as a result
isn't very efficient.  It opens a file for every frame, which due to
memory limitations becomes extremely slow with a lot of frames (SdFat
probably rescans the directory every time), and handling the BMP format
adds overhead.

This mode uses a custom file format which is already swizzled to the
format that we need, so we can read the whole frame in one go, and
uses a single file for a whole animation instead of separate files.
This brings performance from about ~15 FPS (before slowdown) to easily
passing 30 FPS, and videos can be of any length.

Most of the config.ini features are removed.  We don't need to handle
things like scrolling and offsets here.  If you want those, bake them
into the video.  Videos are named "xxx.frm" and are placed in the root
directory, instead of a directory per animation.

The convert.py converts from an image sequence (or a single image) to
this file format.  It doesn't currently handle any settings in
config.ini and the framerate is hardcoded to 30 FPS.